Test enhancements
